WEB-3460: separate queue module into consumer/producer modules
