+++ src/app.module.ts
@@ -15,9 +15,9 @@ import { RacesModule } from './races/races.module'
 import { JwtAuthStrategy } from './authentication/auth-strategies/JwtAuth.strategy'
 import { JwtAuthGuard } from './authentication/guards/JwtAuth.guard'
 import { AdminModule } from './admin/admin.module'
-import { QueueModule } from './queue/queue.module'
+import { QueueConsumerModule } from './queue/consumer/consumer.module'
 import { TopIssuesModule } from './topIssues/topIssues.module'
-import { SharedModule } from './shared/shared.module';
+import { SharedModule } from './shared/shared.module'
 
 @Module({
   imports: [
@@ -33,7 +33,7 @@ import { SharedModule } from './shared/shared.module';
     RacesModule,
     TopIssuesModule,
     AdminModule,
-    QueueModule,
+    QueueConsumerModule,
     SharedModule,
   ],
   controllers: [AppController],
